Rename the Regress integration tests
